Normative: Add text/numeric modes
To support the new mode, in particular cases like anteayer in Spanish, where
there's no dual PluralRule, the locale database is extended with numbered
values, which are queried just with type: "text", before falling back to
the numeric style. This looks like the logic that would be expected based
on the CLDR schema.