Normative: Use the unit in the part description
Rather than calling the numerical part of the result type: "number", it
will have a name which matches the unit--this will make the output
more analogous to DateTimeFormat.prototype.formatToParts.

Additionally, this patch ensures that a literal string like "today"
is wrapped in the appropriate type to make it analogous to patterns.
Such strings will have type: "literal".

Closes #30