Editorial: Simplify parameters to internal constructor calls
RelativeTimeFormat previously passed PluralRules an options bag as an
argument: The type selected was cardinal, which was the default, and
therefore unnecessary and redundant to pass in. This patch removes
the options bag argument.

This patch also makes the minor editorial simplification to pass
the locale directly as an argument to the PluralRules and NumberFormat
constructors, rather than wrapping it in an Array (both have the same
behavior).